[wip] single dataset validation
