+++ bin/check-input-datasets.js
@@ -1,4 +1,5 @@
 const fsPromises = require("fs").promises;
+const path = require("path");
 const {
   getInputDatasetSchema,
 } = require("../src/data-validation/get-input-dataset-schema");
@@ -6,77 +7,19 @@ const { DATA_FOLDER_NAME } = require("../src/process-single-dataset/constants");
 const utils = require("../src/utils");
 const dataPrep = require("../src/data-validation/data-prep");
 const unpackInputDataset = require("../src/data-validation/unpack-input-dataset");
+const {checkForError, checkSingleDatasetInput, datasetCheck} = require("./check-single-dataset");
 // referenced partial schemas
 const INPUT_DATASET_SCHEMA_FILE = "input-dataset.schema.json";
 const INPUT_MEGASET_SCHEMA_FILE = "input-megaset.schema.json";
 const ajv = getInputDatasetSchema();
 
-const checkForError = (fileName, json, schemaFileName) => {
-  const { valid, error } = dataPrep.validate(
-    json,
-    ajv.getSchema(schemaFileName)
-  );
-  let datasetsError = false;
-  let ErrorMsg = "";
-  if (json["feature-defs"]) {
-    const featuresDataOrder = json.dataset.featuresDataOrder;
-    const featureDefsData = json["feature-defs"];
-    const result = utils.validateFeatureDataKeys( featuresDataOrder, featureDefsData );
-    if (result.featureKeysError) {
-      datasetsError = result.featureKeysError;
-      ErrorMsg = result.keysErrorMsg;
-    };
-  };
-  if (json["dataset"]) {
-    const totalCells = json.dataset.userData.totalCells;
-    const totalFOVs = json.dataset.userData.totalFOVs;
-    const result = utils.validateUserDataValues( totalCells, totalFOVs );
-    if (result.userDataError) {
-      datasetsError = result.userDataError;
-      ErrorMsg = result.userDataErrorMsg;
-    };  
-  };
 
-
-  if (!valid || datasetsError) {
-    console.log(
-      "\x1b[0m",
-      `${fileName}`,
-      "\x1b[31m",
-      `failed because: ${JSON.stringify(error || ErrorMsg)}`,
-      "\x1b[0m"
-    );
-    return true;
-  } else {
-    console.log(
-      "\x1b[0m",
-      `${fileName}: check against ${schemaFileName}`,
-      "\x1b[32m",
-      "passed",
-      "\x1b[0m"
-    );
-    return false;
-  }
-};
-
-const validateDatasets = () => {
+const validateDatasets = (singleDatasetFolder = null) => {
   fsPromises
     .readdir(`./${DATA_FOLDER_NAME}`)
     .then(async (files) => {
       const foldersToCheck = [];
       let hasError = false;
-      const checkSingleDatasetInput = async (datasetFolder) => {
-        const inputDataset = await unpackInputDataset(datasetFolder);
-        const foundError = checkForError(
-          `${datasetFolder}`,
-          inputDataset,
-          INPUT_DATASET_SCHEMA_FILE
-        );
-        if (foundError) {
-          hasError = true;
-        }
-      };
-
       for (const name of files) {
         try {
           const subFiles = await fsPromises.readdir(
@@ -90,6 +33,13 @@ const validateDatasets = () => {
         }
       }
       for (const datasetFolder of foldersToCheck) {
+        // skip this dataset if singleDatasetFolder is specified
+        if (singleDatasetFolder && path.basename(datasetFolder) !== path.basename(singleDatasetFolder)) { 
+          continue; 
+        } else if (singleDatasetFolder) {
+          console.log(`logging featureName and order... ${datasetFolder}`);
+          datasetCheck();
+        }; //log feature info, maybe somewhere else?
         const topLevelJson = await utils.readDatasetJson(datasetFolder);
         if (topLevelJson.datasets) {
           // is a megaset, need to check both megaset
@@ -104,10 +54,16 @@ const validateDatasets = () => {
           }
           for (const subDatasetFolder of topLevelJson.datasets) {
             const datasetReadFolder = `${datasetFolder}/${subDatasetFolder}`;
-            await checkSingleDatasetInput(datasetReadFolder);
+            const foundSubError = await checkSingleDatasetInput(datasetReadFolder);
+            if (foundSubError) {
+              hasError = true;
+            }
           }
         } else {
-          await checkSingleDatasetInput(datasetFolder);
+          const foundError = await checkSingleDatasetInput(datasetFolder);
+          if (foundError) {
+            hasError = true;
+          }
         }
       }
       return hasError;
@@ -120,4 +76,8 @@ const validateDatasets = () => {
     });
 };
 
-validateDatasets();
+if (process.argv[2]) {
+  validateDatasets(process.argv[2]);
+} else {
+  validateDatasets();
+}
