+++ package.json
@@ -17,7 +17,8 @@
     "prepublish-docs": "make docs",
     "publish-docs": "gh-pages -d dist",
     "postpublish-docs": "cd src/data-validation && make clean",
-    "validate-datasets": "node bin/check-input-datasets.js"
+    "validate-datasets": "node bin/check-input-datasets.js",
+    "validate-single-dataset": "node bin/check-input-datasets.js"
   },
   "repository": {
     "type": "git",
